+++ dutchiepay/src/app/_components/_user/_login/SocialLogin.jsx
@@ -3,30 +3,18 @@
 import '@/styles/globals.css';
 import '@/styles/user.css';
 
-import { useEffect, useState } from 'react';
-
 import Cookies from 'universal-cookie';
 import CryptoJS from 'crypto-js';
-import Image from 'next/image';
-import LastLogin from './LastLogin';
-import kakao from '../../../../public/image/kakao.png';
+import SocialButton from './SocialButton';
 import { login } from '@/redux/slice/loginSlice';
-import naver from '../../../../public/image/naver.png';
 import { useDispatch } from 'react-redux';
+import { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 
 export default function SocialLogin() {
   const router = useRouter();
   const cookies = new Cookies();
   const dispatch = useDispatch();
-
-  const openPopup = (type) => {
-    window.open(
-      `${process.env.NEXT_PUBLIC_BASE_URL}/oauth/signup?type=${type}`,
-      `${type} 회원가입`,
-      'width=600,height=400'
-    );
-  };
 
   useEffect(() => {
     const handleMessage = (event) => {
@@ -89,22 +77,8 @@ export default function SocialLogin() {
     <section>
       <h2 className="user-login-sns__title">&nbsp;SNS 계정으로 로그인&nbsp;</h2>
       <div className="flex mt-[24px] gap-[60px] items-center justify-center">
-        <button className="relative" onClick={() => openPopup('naver')}>
-          <Image
-            className="w-[60px] h-[60px] cursor-pointer"
-            src={naver}
-            alt="naver"
-          />
-          <LastLogin type={'naver'} />
-        </button>
-        <button className="relative" onClick={() => openPopup('kakao')}>
-          <Image
-            className="w-[60px] h-[60px] cursor-pointer"
-            src={kakao}
-            alt="kakao"
-          />
-          <LastLogin type={'kakao'} />
-        </button>
+        <SocialButton type={'naver'} />
+        <SocialButton type={'kakao'} />
       </div>
     </section>
   );
